refactor: restructure SectionHeader component and remove outdated stories
